support leader complain in block version 0.1a.
