+++ loopchain/blockchain/blocks/block_verifier.py
@@ -11,12 +11,14 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
 import hashlib
 from abc import ABC, abstractmethod
 from typing import TYPE_CHECKING, Callable
 
 from secp256k1 import PrivateKey, PublicKey
 
+from loopchain import utils
 from .. import ExternalAddress, BlockVersionNotMatch
 
 if TYPE_CHECKING:
@@ -60,6 +62,16 @@ class BlockVerifier(ABC):
         if expect_address != block.header.peer_id:
             raise RuntimeError(f"block peer id {block.header.peer_id.hex_xx()}, "
                                f"expected {ExternalAddress(expect_address).hex_xx()}")
+
+    def verify_generator(self, block: 'Block', generator: 'ExternalAddress'):
+        utils.logger.notice(f"verify_generator \nheight({block.header.height}) "
+                            f"\ngenerator({block.header.peer_id.hex_xx()}) "
+                            f"\nexpected({generator.hex_xx()}) "
+                            f"\nnext_leader({block.header.next_leader.hex_xx()})")
+        if not block.header.complained and block.header.peer_id != generator:
+            raise RuntimeError(f"Block({block.header.height}, {block.header.hash.hex()}, "
+                               f"Generator({block.header.peer_id.hex_xx()}), "
+                               f"Expected({generator.hex_xx()}).")
 
     @classmethod
     def new(cls, version: str, tx_versioner: 'TransactionVersioner') -> 'BlockVerifier':
