+++ loopchain/blockchain/blocks/v0_1a/block.py
@@ -1,5 +1,7 @@
 from collections import OrderedDict
 from dataclasses import dataclass
+
+from loopchain import utils
 from .. import BlockHeader as BaseBlockHeader, BlockBody as BaseBlockBody, _dict__str__
 from ... import Hash32, Address, Signature, ExternalAddress
 
@@ -25,6 +27,20 @@ class BlockHeader(BaseBlockHeader):
             commit_state = OrderedDict(commit_state)
             commit_state.__str__ = _dict__str__
             object.__setattr__(self, "commit_state", commit_state)
+
+    @property
+    def complained(self):
+        """Version compatible in (v0.1a, v0.2)
+
+        :return:
+        """
+        # tx == 0 and peer_id == next_leader >> complained = True
+        complained = self.peer_id == self.next_leader and self.merkle_tree_root_hash == Hash32(bytes(32))
+        utils.logger.notice(f"get v0.1 complained \nheight({self.height}) "
+                            f"\nmerkle_tree_root_hash({self.merkle_tree_root_hash})"
+                            f"\ncomplained({complained})")
+
+        return complained
 
 
 @dataclass(frozen=True)
